// src/app/generated/model/documentDetailDTO.ts
/**
 * document-management
 * 1000kit document management
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryDTO } from './categoryDTO';
import { AttachmentDTO } from './attachmentDTO';
import { RelatedObjectRefDTO } from './relatedObjectRefDTO';
import { ChannelDTO } from './channelDTO';
import { LifeCycleState } from './lifeCycleState';
import { DocumentCharacteristicDTO } from './documentCharacteristicDTO';
import { DocumentSpecificationDTO } from './documentSpecificationDTO';
import { RelatedPartyRefDTO } from './relatedPartyRefDTO';
import { DocumentTypeDTO } from './documentTypeDTO';
import { DocumentRelationshipDTO } from './documentRelationshipDTO';

export interface DocumentDetailDTO {
  creationDate?: string;
  creationUser?: string;
  modificationDate?: string;
  modificationUser?: string;
  version?: number;
  id?: string;
  attachments?: Array<AttachmentDTO>;
  categories?: Array<CategoryDTO>;
  channel?: ChannelDTO;
  characteristics?: Array<DocumentCharacteristicDTO>;
  description?: string;
  documentRelationships?: Array<DocumentRelationshipDTO>;
  documentVersion?: string;
  lifeCycleState?: LifeCycleState;
  name?: string;
  relatedObject?: RelatedObjectRefDTO;
  relatedParties?: Array<RelatedPartyRefDTO>;
  specification?: DocumentSpecificationDTO;
  tags?: Array<string>;
  type?: DocumentTypeDTO;
}